:recycle: refactor(notice): Move ResponseDto to package
