+++ src/main/java/nbc/mushroom/domain/like/repository/AuctionItemLikeRepositoryCustom.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Optional;
 import nbc.mushroom.domain.auction_item.entity.AuctionItem;
 import nbc.mushroom.domain.like.entity.AuctionItemLike;
-import nbc.mushroom.domain.notice.dto.NoticeRes;
+import nbc.mushroom.domain.notice.dto.response.NoticeRes;
 import nbc.mushroom.domain.user.dto.response.SearchUserAuctionItemLikeRes;
 import nbc.mushroom.domain.user.entity.User;
 import org.springframework.data.domain.PageImpl;
