+++ src/main/java/nbc/mushroom/domain/notice/service/NoticeReadService.java
@@ -3,7 +3,7 @@ package nbc.mushroom.domain.notice.service;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nbc.mushroom.domain.notice.dto.ReadNoticeRes;
+import nbc.mushroom.domain.notice.dto.response.ReadNoticeRes;
 import nbc.mushroom.domain.notice.entity.Notice;
 import nbc.mushroom.domain.notice.repository.NoticeRepository;
 import nbc.mushroom.domain.user.entity.User;
